requirements.txt
asgiref==3.5.2
Django==4.1.1
django-extensions==3.2.1
psycopg2==2.9.3
python-decouple==3.6
sqlparse==0.4.3
